requirements.txt
pytest==7.4.0
PyGithub==1.59.0
